backend/exam/pyproject.toml
[tool.poetry]
name = "online-exam"
version = "0.1.0"
description = "考试系统"
authors = ["xingxingzaixian <beixia1989@163.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "3.2.4"
djangorestframework = "3.12.4"
django-cors-headers = "3.7.0"
python-jose = "3.3.0"
PyMySQL = "1.0.2"
django-filter = "2.4.0"
coreapi = "2.3.3"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
dev = "python manage.py runserver 0.0.0.0:8888"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
